Move server implementation to tchannel.thrift.server
