+++ python/tchannel/thrift/__init__.py
@@ -20,98 +20,12 @@
 
 from __future__ import absolute_import
 
-import inspect
-
-from thrift.transport import TTransport
-from thrift.protocol import TBinaryProtocol
-
-from .protocol import TChannelProtocolFactory, TChannelProtocol  # noqa
+from .server import register  # noqa
+from .protocol import TChannelProtocol  # noqa
+from .protocol import TChannelProtocolFactory  # noqa
 from .transport import TChannelTransport  # noqa
+
 try:
     from .transport.tornado import TChannelTornadoTransport  # noqa
 except ImportError:
     pass
-
-
-def register(dispatcher, service_module, handler, service_name=None):
-    """Registers the given service with the given dispatcher.
-
-    .. code-block:: python
-
-        # For,
-        #
-        #   service HelloWorld { string hello(1: string name); }
-
-        import tchannel.thrift
-        import HelloWorld
-
-        class HelloWorldHandler(object):
-            def hello(self, name):
-                return "Hello, %s" % name
-
-        dispatcher = TornadoDispatcher()
-        tchannel.thrift.register(dispatcher, HelloWorld, HelloWorldHandler())
-
-    :param dispatcher:
-        TChannel dispatcher with which the Thrift service will be registered.
-    :param service_module:
-        The service module generated by Thrift. This module contains the
-        service ``Iface``, ``Client``, ``Processor``, etc. classes.
-    :param handler:
-        An object implementing the service interface.
-    :param service_name:
-        Thrift service name. This is the `service` name specified in the
-        Thrift IDL. If omitted, it is automatically determined based on the
-        name of ``service_module``.
-    """
-    assert handler, "hanlder can't be None"
-    assert dispatcher, "dispatcher can't be None"
-    assert service_module, "service_module can't be None"
-
-    if not service_name:
-        service_name = service_module.__name__.rsplit('.', 1)[-1]
-
-    methods = [
-        name for (name, _) in inspect.getmembers(
-            service_module.Iface, predicate=inspect.ismethod
-        )
-    ]
-
-    for method in methods:
-        endpoint = "%s::%s" % (service_name, method)
-        dispatcher.register(
-            endpoint, build_handler(service_module, method, handler)
-        )
-
-
-def build_handler(service_module, method_name, handler):
-    def thrift_handler(request, response, connection):
-        # TODO: Fix arg scheme passing
-        # assert request.arg_scheme == 'thrift', (
-        #     "Invalid arg scheme %s" % request.arg_scheme
-        # )
-
-        args = getattr(service_module, method_name + '_args')()
-        args.read(
-            TBinaryProtocol.TBinaryProtocolAccelerated(
-                TTransport.TMemoryBuffer(request.body)
-            )
-        )
-
-        result = getattr(service_module, method_name + '_result')()
-
-        args = [getattr(args, spec[2]) for spec in args.thrift_spec[1:]]
-        try:
-            result.success = getattr(handler, method_name)(*args)
-        except Exception as exc:
-            for spec in result.thrift_spec[1:]:
-                if isinstance(exc, spec[3][0]):
-                    setattr(result, spec[2], exc)
-                    break
-            else:
-                raise
-
-        output_buf = TTransport.TMemoryBuffer()
-        result.write(TBinaryProtocol.TBinaryProtocolAccelerated(output_buf))
-        response.write(arg3=output_buf.getvalue())
-    return thrift_handler
