+++ python/tchannel/thrift/server.py
@@ -20,51 +20,96 @@
 
 from __future__ import absolute_import
 
-from ..handler import RequestHandler
+import inspect
 
-# Ignore this file for now. This is just a placeholder to add all the
-# server-side support.
-
-
-def processor(thrift_module, implementation_class, exception_logger=None):
-    # TODO probably lift from clay_thrift/common.py
-    raise NotImplementedError()
+from thrift.transport import TTransport
+from thrift.protocol import TBinaryProtocol
 
 
-class ThriftDispatcher(RequestHandler):
-    """A RequestHandler that delegates calls to a Thrift processor.
+def register(dispatcher, service_module, handler, service_name=None):
+    """Registers the given service with the given dispatcher.
 
-    .. code-block::
+    .. code-block:: python
 
-        dispatcher = ThriftDispatcher()
+        # For,
+        #
+        #   service HelloWorld { string hello(1: string name); }
 
+        import tchannel.thrift
+        import HelloWorld
+
+        class HelloWorldHandler(object):
+            def hello(self, name):
+                return "Hello, %s" % name
+
+        dispatcher = TornadoDispatcher()
+        tchannel.thrift.register(dispatcher, HelloWorld, HelloWorldHandler())
+
+    :param dispatcher:
+        TChannel dispatcher with which the Thrift service will be registered.
+    :param service_module:
+        The service module generated by Thrift. This module contains the
+        service ``Iface``, ``Client``, ``Processor``, etc. classes.
+    :param handler:
+        An object implementing the service interface.
+    :param service_name:
+        Thrift service name. This is the `service` name specified in the
+        Thrift IDL. If omitted, it is automatically determined based on the
+        name of ``service_module``.
+    """
+    assert handler, "hanlder can't be None"
+    assert dispatcher, "dispatcher can't be None"
+    assert service_module, "service_module can't be None"
+
+    if not service_name:
+        service_name = service_module.__name__.rsplit('.', 1)[-1]
+
+    methods = [
+        name for (name, _) in inspect.getmembers(
+            service_module.Iface, predicate=inspect.ismethod
+        )
+    ]
+
+    for method in methods:
+        endpoint = "%s::%s" % (service_name, method)
         dispatcher.register(
-            "UserService",
-            processor(
-                blog_service.thrift.service.UserService,
-                UserServiceHandler,
-            ),
+            endpoint, build_handler(service_module, method, handler)
         )
 
-        dispatcher.register(
-            "PostService",
-            processor(
-                blog_service.thrift.service.PostService,
-                PostServiceHandler,
+
+def build_handler(service_module, method_name, handler):
+    args_type = getattr(service_module, method_name + '_args')
+    result_type = getattr(service_module, method_name + '_result')
+
+    def thrift_handler(request, response, connection):
+        # TODO: Fix arg scheme passing
+        # assert request.arg_scheme == 'thrift', (
+        #     "Invalid arg scheme %s" % request.arg_scheme
+        # )
+
+        args = args_type()
+        args.read(
+            TBinaryProtocol.TBinaryProtocolAccelerated(
+                TTransport.TMemoryBuffer(request.body)
             )
         )
 
+        result = result_type()
+        try:
+            result.success = getattr(handler, method_name)(
+                *[getattr(args, spec[2]) for spec in args.thrift_spec[1:]]
+            )
+        except Exception as exc:
+            for spec in result.thrift_spec[1:]:
+                if isinstance(exc, spec[3][0]):
+                    setattr(result, spec[2], exc)
+                    break
+            else:
+                raise
 
-    """
+        output_buf = TTransport.TMemoryBuffer()
+        result.write(TBinaryProtocol.TBinaryProtocolAccelerated(output_buf))
 
-    __slots__ = ('services',)
+        response.write(arg3=output_buf.getvalue())
 
-    def __init__(self):
-        super(ThriftDispatcher, self).__init__()
-        self.services = {}
-
-    def register(self, service_name, processor):
-        self.services[service_name] = processor
-
-    def handle(self, context, connection):
-        raise NotImplementedError()
+    return thrift_handler
