+++ python/tox.ini
@@ -15,4 +15,5 @@ commands =
   py.test --cov tchannel --cov-report=xml --cov-report=term-missing {posargs}
 
 [flake8]
-exclude = examples/*
+# Ignore files generated by Thrift
+exclude = examples/thrift/hello/*
